+++ sketch-core/src/androidInstrumentedTest/kotlin/com/github/panpf/sketch/core/android/test/AndroidBitmapImageTest.kt
@@ -24,7 +24,6 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotEquals
-import kotlin.test.assertNotNull
 import kotlin.test.assertSame
 import kotlin.test.assertTrue
 
@@ -118,15 +117,6 @@ class AndroidBitmapImageTest {
     fun testTransformer() {
         AndroidBitmapImage(AndroidBitmap(100, 200)).apply {
             assertTrue(actual = transformer() is AndroidBitmapImageTransformer)
-        }
-    }
-
-    @Test
-    fun testGetPixels() {
-        AndroidBitmapImage(AndroidBitmap(100, 200)).apply {
-            val pixels = getPixels()
-            assertNotNull(actual = pixels)
-            assertEquals(expected = 20000, actual = pixels.size)
         }
     }