+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Translations: [简体中文](CHANGELOG_zh.md)
 * fix: Fix the bug that the image cannot be loaded when the ImageView is attached to the window but
   the size is null due to padding. [#208](https://github.com/panpf/sketch/issues/208)
 * remove: Remove ComposeBitmapImage
+* remove: Remove Image.getPixels()
 * change: SkiaBitmapImage is now cached in the memory cache on non-Android platforms, not
   ComposeBitmapImage.
 * change: The parameter passed in Fetcher.Factory.create() is changed to RequestContext